+++ src/app/search/page.tsx
@@ -10,6 +10,7 @@ import { useRoomMeta } from "@/src/entities/room/hooks/useRoomMeta";
 import { getDefaultRoomImage } from "@/src/entities/room/lib/getDefaultRoomImage";
 import { useRoomNavigator } from "@/src/shared/lib/useRoomNavigator";
 import { useLoadMoreRoomsNearEnd } from "@/src/shared/lib/useLoadMoreRoomsNearEnd";
+import { useAuthenticatedAction } from "@/src/shared/lib/useAuthenticatedAction";
 import { SearchPageRoomList } from "@/src/features/room/search/ui/SearchPageRoomList";
 import MainLogo from "@/src/widgets/home/ui/MainLogo";
 import styles from "./page.module.css";
@@ -28,20 +29,20 @@ import { useRoomEntry } from "@/src/features/room/join/model/useRoomEntry";
 import RoomJoinPasswordModal from "@/src/features/room/join/ui/RoomJoinPasswordModal";
 import FollowModal from "@/src/features/follow/ui/FollowModal";
 import SettingsModal from "@/src/features/settings/ui/SettingsModal";
-import { useMe } from "@/src/entities/user/hooks/useMe";
 import { redirectToGoogleLogin } from "@/src/features/auth/login-with-google/api/login";
 import AuthRequiredModal from "@/src/shared/ui/auth-required/AuthRequiredModal";
 
 export default function SearchPage() {
   const [roomListFilters, setRoomListFilters] = useState(DEFAULT_HOME_FILTERS);
   const [isCreateRoomModalOpen, setIsCreateRoomModalOpen] = useState(false);
-  const [isAuthRequiredModalOpen, setIsAuthRequiredModalOpen] = useState(false);
-  const [authRequiredDescription, setAuthRequiredDescription] = useState(
-    "방 만들기는 로그인 후 이용할 수 있어요.",
-  );
   const [isFollowModalOpen, setIsFollowModalOpen] = useState(false);
   const [isSettingsModalOpen, setIsSettingsModalOpen] = useState(false);
-  const { data: me, refetch: refetchMe } = useMe();
+  const {
+    authRequiredDescription,
+    closeAuthRequiredModal,
+    isAuthRequiredModalOpen,
+    requestAuthenticatedAction,
+  } = useAuthenticatedAction("방 만들기는 로그인 후 이용할 수 있어요.");
   const {
     data,
     fetchNextPage,
@@ -95,37 +96,23 @@ export default function SearchPage() {
     );
   };
 
-  const requestAuthenticatedAction = async (
-    onAuthenticated: () => void,
-    description: string,
-  ) => {
-    if (me) {
-      onAuthenticated();
-      return;
-    }
-
-    const result = await refetchMe();
-
-    if (result.data) {
-      onAuthenticated();
-      return;
-    }
-
-    setAuthRequiredDescription(description);
-    setIsAuthRequiredModalOpen(true);
-  };
-
   const requestCreateRoom = () =>
-    requestAuthenticatedAction(
-      () => setIsCreateRoomModalOpen(true),
-      "방 만들기는 로그인 후 이용할 수 있어요.",
-    );
+    requestAuthenticatedAction({
+      description: "방 만들기는 로그인 후 이용할 수 있어요.",
+      onAuthenticated: () => setIsCreateRoomModalOpen(true),
+    });
 
   const requestOpenFollow = () =>
-    requestAuthenticatedAction(
-      () => setIsFollowModalOpen(true),
-      "친구 기능은 로그인 후 이용할 수 있어요.",
-    );
+    requestAuthenticatedAction({
+      description: "친구 기능은 로그인 후 이용할 수 있어요.",
+      onAuthenticated: () => setIsFollowModalOpen(true),
+    });
+
+  const requestOpenSettings = () =>
+    requestAuthenticatedAction({
+      description: "설정은 로그인 후 이용할 수 있어요.",
+      onAuthenticated: () => setIsSettingsModalOpen(true),
+    });
 
   return (
     <div className={styles.container}>
@@ -230,7 +217,7 @@ export default function SearchPage() {
           onSelectFilter={selectRoomListFilter}
           onCreateRoom={requestCreateRoom}
           onOpenFollow={requestOpenFollow}
-          onOpenSettings={() => setIsSettingsModalOpen(true)}
+          onOpenSettings={requestOpenSettings}
           onEnterSelectedRoom={() => {
             if (selectedRoom) {
               roomEntry.requestRoomEntry(selectedRoom);
@@ -262,7 +249,7 @@ export default function SearchPage() {
       <AuthRequiredModal
         open={isAuthRequiredModalOpen}
         description={authRequiredDescription}
-        onClose={() => setIsAuthRequiredModalOpen(false)}
+        onClose={closeAuthRequiredModal}
         onLogin={redirectToGoogleLogin}
       />
     </div>
