+++ src/features/auth/login-with-google/ui/LoginModal.module.css
@@ -1,7 +1,7 @@
 .modalOverlay {
   position: fixed;
   inset: 0;
-  z-index: 1000;
+  z-index: 1200;
   display: flex;
   align-items: center;
   justify-content: center;
