+++ src/features/follow/ui/FollowModal.module.css
@@ -1,12 +1,11 @@
 .overlay {
   position: fixed;
   inset: 0;
-  z-index: 1000;
+  z-index: 1200;
   display: flex;
   align-items: center;
   justify-content: center;
   padding: 24px;
-  bottom: 130px;
 }
 
 .modal {
