+++ src/features/room/create/ui/RoomFormModal.module.css
@@ -1,8 +1,7 @@
 .overlay {
   position: fixed;
   inset: 0;
-  z-index: 1000;
-  bottom: 130px;
+  z-index: 1200;
   display: flex;
   align-items: center;
   justify-content: center;
